+++ src/Stormancer.Plugins/Epic/Stormancer.Server.Plugins.Epic/Changelog.rst
@@ -9,6 +9,12 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+Unreleased
+----------
+Changed
+*******
+- Use latest version of the MemoryCache
+
 1.1.3
 -----
 Fixed